refactor: extract AppState system-actions extension (#572, #430)
Second slice of #430 (AppState decomposition), following the extension-
per-domain pattern established by 430a (#571). Extracts the 4 privacy-
settings + update-check delegator methods (openMicrophonePrivacySettings,
openScreenRecordingPrivacySettings, openSystemAudioPrivacySettings,
checkForUpdates).

Moves AppState.swift lines 894-908 (15 lines, 555 bytes, SHA-256
e2c94461244733c1d30a17a4fcc6d6e369d70498094e09abf64ecb0f8f95129b) into a
new Sources/BugNarrator/AppState+SystemActions.swift as `extension
AppState { ... }`. Each method is a single-call delegator through
appUtilityActionPresenter.present(appUtilityActions.<method>()) — same
shape as 4 of the 7 methods in 430a. Byte-identical bodies.

Endpoints verified before extraction (sed 894p = `func
openMicrophonePrivacySettings() {`, 908p = `}` of checkForUpdates).
Byte-diff against origin/main lines 894-908 is silent (exit 0); SHA-256
of the extracted block in the new file matches the origin SHA.

The extension does NOT re-annotate @MainActor — inherited from
AppState's class-level annotation per SE-0316 (proven by 430a).
Foundation-only import matches house style. 662 unit tests pass
unchanged; docs-sync validation passes.

Pre-build spec reviewed by codex in a fresh session (3 findings folded
in — most important was a range off-by-one catch: my initial ticket
claimed 894-909/16 lines but the actual body is 894-908/15 lines; line
909 was the trailing separator blank. codex-verified endpoints + SHA
provided the corrected values before build).
